# specs/001-sports-capture-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CourtVision Platform API
  version: 0.1.0
  description: Public QR/player, owner, capture-agent, and delivery boundaries for the MVP.
servers:
  - url: https://api.example.invalid
...
paths:
...
  /agent/heartbeat:
    post:
      tags: [Capture Agent]
      summary: Report camera agent health
      security:
        - agentAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeartbeatRequest'
      responses:
        '204':
          description: Health accepted.
  /agent/sessions/{sessionId}/events:
    post:
      tags: [Capture Agent]
      summary: Submit an idempotent normalized highlight event
      security:
        - agentAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureEventRequest'
      responses:
        '202':
          description: Event accepted for clip processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventAcceptedResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /agent/uploads/presign:
    post:
      tags: [Capture Agent]
      summary: Request a short-lived upload target for a source segment or recording
      security:
        - agentAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignUploadRequest'
      responses:
        '200':
          description: Upload target and metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignUploadResponse'
...
components:
  securitySchemes:
    ownerAuth:
      type: http
      scheme: bearer
    agentAuth:
      type: http
      scheme: bearer
    accessToken:
      type: http
      scheme: bearer
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8 }
  responses:
...
    Conflict:
      description: Duplicate or conflicting active resource.
...
  schemas:
...
    HeartbeatRequest:
      type: object
      required: [device_id, observed_at, camera_status, agent_version]
      properties:
        device_id: { type: string }
        observed_at: { type: string, format: date-time }
        camera_status: { type: string, enum: [online, offline, degraded] }
        active_session_id: { type: string, format: uuid, nullable: true }
        agent_version: { type: string }
    CaptureEventRequest:
      type: object
      required: [source_id, event_type, occurred_at]
      properties:
        source_id: { type: string }
        event_type: { type: string, enum: [gesture, manual, physical_button, automatic_sport_event] }
        occurred_at: { type: string, format: date-time }
        confidence: { type: number, minimum: 0, maximum: 1, nullable: true }
        buffer_start_at: { type: string, format: date-time }
        buffer_end_at: { type: string, format: date-time }
        initiating_player_id: { type: string, format: uuid, nullable: true }
        source_storage_key: { type: string, nullable: true }
    EventAcceptedResponse:
      type: object
      required: [event_id, highlight_id, status]
      properties:
        event_id: { type: string, format: uuid }
        highlight_id: { type: string, format: uuid }
        status: { type: string, enum: [accepted, processing] }
    PresignUploadRequest:
      type: object
      required: [session_id, media_type, content_type, size_bytes, checksum]
      properties:
        session_id: { type: string, format: uuid }
        media_type: { type: string, enum: [segment, recording, highlight_source] }
        content_type: { type: string }
        size_bytes: { type: integer, minimum: 1 }
        checksum: { type: string }
    PresignUploadResponse:
      type: object
      required: [upload_url, storage_key, expires_at]
      properties:
        upload_url: { type: string, format: uri }
        storage_key: { type: string }
        expires_at: { type: string, format: date-time }
...